include confirm delete

+++ src/components/course/CourseList.js
@@ -5,28 +5,46 @@ import Swal from "sweetalert2";
 
 export default function CourseList() {
   const [students, setStudents] = useState([]);
+  const [error, setError] = useState(null);
 
-   async function deletdStudent (e, id) {
+  function deletdStudent(e, id) {
     e.preventDefault();
+    try {
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          fetch(
+            `https://www.full-stack-app.com/services/public/api/delete-student/${id}`,
+            { method: "DELETE" }
+          )
+            .then((res) => res.json())
+            .then((res) => console.log(res.student));
 
-    await fetch(
-      `https://www.full-stack-app.com/services/public/api/delete-student/${id}`,
-      { method: "DELETE" }
-    )
-      .then((res) => res.json())
-      .then((res)=>console.log(res.student));
-
-  //  Swal.fire({
-  //     position: "center",
-  //     icon: "success",
-  //     title: "Your work has been deleted",
-  //     showConfirmButton: false,
-  //     timer: 2500,
-  //   });
-
-    window.location.reload(0)
-
-  };
+            Swal.fire({
+              title: 'Deleted!',
+              text : 'Your file has been deleted.',
+              icon: 'success',
+              showConfirmButton: false,
+              timer: 1500
+            })
+          setTimeout(function () {
+            window.location.reload();
+          }, 1500);
+        } else {
+          console.log(error)
+        }
+      });
+    } catch (error) {
+      setError(error);
+    }
+  }
 
   const fetchData = () => {
     //fetch("http://localhost/laravel-react-crud/public/api/students")
